BaseAtGoodStage
The base is at a pretty good stage. I am going to begin creating a survivalIsland scene and adding some specific functionality. This will be done without making many changes to the base that will need removing before using in another project. Lots of work needs doing to neaten up the code and improve efficiency however the decision is that can wait for the moment.