Beginning of behaviour nodes
Currently the project is at a good stage to save to the branch before getting complex. New features of note are the beginning of behaviour nodes though further planning is required. And most importantly coloured debug messages for the unity console. While I need to put more effort into planning the color formatting to make it look nice it is at a good enough stage for the moment.